pyproject.toml
[project]
name = "claim-api"
version = "0.1.0"
description = "Stateless API for extracting topic-organized claims from source text using Gemini structured outputs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "google-genai>=1.50.0",
    "json-repair>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
